requirements.txt
Django>=3.1.7,<3.2.0
pylint>=2.7.4,<2.8.0
pylint-django>=2.4.2,<2.5.0
djangorestframework>=3.12.4,<3.13.0
Markdown>=3.3.4,<3.4.0
django-filter>=2.4.0,<3.4.0
psycopg2-binary>=2.8.6,<2.9.0
django-environ>=0.4.5,<1.0.0
asgiref>=3.3.1,<3.4.0
validate-docbr>=1.8.2,<1.9.0
Faker>=7.0.1,<7.1.0
six>=1.15.0,<1.16.0
sqlparse>=0.4.1,<0.5.0
pytz>=2021.1,<2021.2
text-unidecode>=1.3,<1.4.0
python-dateutil>=2.8.1,<2.9.0